requirements.txt
arxiv==2.1.0
PyYAML==6.0.2
Jinja2==3.1.4
markdown==3.6
openai>=1.30.6
tenacity==8.4.2
python-dateutil==2.9.0.post0
requests==2.32.3
